работаем

+++ src/data/Coordinates.java
@@ -1,17 +1,17 @@
 package data;
 
 public class Coordinates {
-    private long x;
-    private Integer y; //Поле не может быть null
+    private Float x;
+    private float y; //Поле не может быть null
 
-    public Coordinates(long x, Integer y){
+    public Coordinates(Float x, float y){
         this.x = x;
         this.y = y;
     }
     public Coordinates(){}
 
-    public long getX(){return x;}
-    public void setX(long x){this.x = x;}
-    public int getY(){return y;}
-    public void setY(Integer y){this.y = y;}
+    public Float getX(){return x;}
+    public void setX(Float x){this.x = x;}
+    public float getY(){return y;}
+    public void setY(float y){this.y = y;}
 }
